refactor Template component to reuse for invoices and estimates
